(ml5717) Added contracts to necsim-core, necsim-classical and rustcoalescence (necsim-impls still missing)

+++ necsim-impls/src/landscape/habitat/mod.rs
@@ -7,7 +7,7 @@ pub trait Habitat {
     fn get_extent(&self) -> LandscapeExtent;
 
     #[must_use]
-    fn get_total_habitat(&self) -> u32;
+    fn get_total_habitat(&self) -> usize;
     #[must_use]
     fn get_habitat_at_location(&self, location: &Location) -> u32;
 }
